rebirth

+++ include/starlia/core/math.h
@@ -27,7 +27,7 @@ double angleOf(Coord2f start, Coord2f end);
 
 double length(Coord2f start, Coord2f end);
 
-double dmod(double first, int second);
+double dmod(double first, double second);
 double clamp(double operand, double dlimit, double ulimit);
 
 Coord2f polarToXY(double radius, double angle);
